fix: make the Windows app compile — shared helpers + BCD cleanup
The Windows build referenced downloadFile/copyFile/marshalJSON that only
existed in the !windows file; move them to a portable helpers.go.
configureBCD now targets the E2E-proven signed-shim chain path and arms
only the one-shot bootsequence (no permanent displayorder change).
